Update README.md
Some of the arguments listed for the Replay tool were renamed and no longer valid as documented. Modified to detail required arguments and to send users to --help to see extended options. Limiting to required options on this page will make it easier to maintain and makes it easier for people to run the first time.