[record] join writer thread after tailer threads
_join_workers uses a simple loop to await worker threads. When ctrl+c is
called, force_quit is set to true, but threads are not killed until
processed by this loop. As the writer thread continues to run until the
queue is empty, this can cause an indefinite wait if _join_workers
blocks on the write thread while tailer threads are still running, since
they continue to write to the queue.

Here we move the join call on the writer thread to the end of the loop.
In addition, stop increasing the the wait_time by a factor of 2 as this
can lead to long and unnecessary waits for the program to shut down.